requirements.txt
lap
shapely
typing-extensions
ultralytics==8.3.128
viam-sdk==0.47.0
torch==2.7.0

